refactor: HandlerMappings와 HandlerAdapters의 초기화 책임 이전
